+++ i18n_flutter/CHANGELOG.md
@@ -3,3 +3,4 @@
 * Added auto-injection of COI Service Worker for web builds.
 * Solve auto-injection of windows version
 * Solve web-injection
+* add android ios, macos injection